+++ starsky/starsky/clientapp/src/components/detailview-gpx.tsx
@@ -1,5 +1,5 @@
 import L from 'leaflet';
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useRef } from 'react';
 import useLocation from '../hooks/use-location';
 import { IConnectionDefault } from '../interfaces/IConnectionDefault';
 import FetchXml from '../shared/fetch-xml';
@@ -9,12 +9,15 @@ import { UrlQuery } from '../shared/url-query';
 const DetailViewGpx: React.FC = () => {
   var history = useLocation();
 
-  const [gpxClassName, setGpxClassName] = useState("main main--error");
-
   function updateMap(response: IConnectionDefault) {
-    console.log(response.data);
-
     if (!response.data) return;
+    if (!mapReference.current) return;
+
+    // reset leaflet first
+    var container = L.DomUtil.get(mapReference.current);
+    if (container != null) {
+      (container as any)._leaflet_id = null;
+    }
 
     var tracks: any[] = [];
     var tracksNodeList: NodeListOf<Element> = (response.data as XMLDocument).querySelectorAll('trkpt');
@@ -24,7 +27,7 @@ const DetailViewGpx: React.FC = () => {
     });
 
     // create map
-    var map = L.map('map', {
+    var map = L.map(mapReference.current, {
       layers: [
         L.tileLayer('http://{s}.tile.osm.org/{z}/{x}/{y}.png', {
           attribution:
@@ -44,18 +47,21 @@ const DetailViewGpx: React.FC = () => {
     map.fitBounds(tracks);
 
     L.polygon(tracks, { color: 'red', fill: false }).addTo(map);
-    console.log('map added');
   }
 
+  // when having to gpx files next and you browse though it
+  const mapReference = useRef<HTMLDivElement>(null);
+
+  /** update only on intial load */
   useEffect(() => {
     var filePathEncoded = new URLPath().encodeURI(new URLPath().getFilePath(history.location.search));
     FetchXml(new UrlQuery().UrlDownloadPhotoApi(filePathEncoded, false)).then((response) => {
       updateMap(response)
     })
-  }, []);
+  }, [new URLPath().getFilePath(history.location.search)]);
 
   return (
-    <div className={"main main--error main--gpx"} id="map" />
+    <div className={"main main--error main--gpx"} ref={mapReference} />
   );
 };